rename `observable_combine_test.dart` to `combine_test.dart`
