+++ test/observables/observables/observable_combine_test.dart
@@ -1,7 +1,7 @@
 
 import 'package:test/test.dart';
 import 'package:scopes/scopes.dart';
-import '../../test_gen/combine_test/observable_combine_test.dart';
+import '../../test_gen/combine_test/combine_test.dart';
 import '../../toolbox/observable_tester.dart';
 
 part 'observable_combine_test.g.dart';
