+++ test/test_gen/test_builder.dart
@@ -2,7 +2,7 @@
 import 'package:build/build.dart';
 import 'package:source_gen/source_gen.dart';
 
-import 'combine_test/observable_combine_test_generator.dart';
+import 'combine_test/combine_test_generator.dart';
 
 Builder testBuilder(BuilderOptions options) {
   return SharedPartBuilder(
